requirements/base.txt
Django==4.2.14
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
psycopg2-binary
python-decouple==3.8
dj-database-url==2.2.0
